intel-xed: add version 12.0.1 (#20676)
* intel-xed: add version 12.0.1

Rework the version numbers for intel-xed, now that xed has actual
releases and tags.  Add releases 11.2.0 and 12.0.1.  Rename 2019.03.01
to 10.2019.03 as a legacy version that fits in the new order.

Add variant +pic to compile libxed.a with PIC code so that it can be
linked into another shared library.

Add conflict for aarch64.
Add mwkrentel as maintainer.